feat: add validate for deleting company only has status inactive

+++ app/Domain/Companies/Actions/DeleteCompany.php
@@ -2,6 +2,7 @@
 
 namespace App\Domain\Companies\Actions;
 
+use App\Constants\CompanyStatus;
 use App\Domain\Companies\Models\Company;
 use Illuminate\Support\Facades\Log;
 
@@ -10,6 +11,10 @@ class DeleteCompany
     public static function execute(Company $company): bool
     {
         try {
+            if($company->status === CompanyStatus::ACTIVE->value){
+                return false;
+            }
+
             return $company->delete();
         } catch (\Exception $e) {
             Log::channel('Company')->error('Error deleting company: '.$e->getMessage());
